[delete][#53] delete Planboard2
